Move ModalContainer into its own file
